Refactor usage permissions handling
