+++ app/src/main/java/com/talauncher/ui/settings/SettingsViewModel.kt
@@ -6,6 +6,7 @@ import com.talauncher.data.model.AppInfo
 import com.talauncher.data.model.InstalledApp
 import com.talauncher.data.repository.AppRepository
 import com.talauncher.data.repository.SettingsRepository
+import com.talauncher.utils.PermissionsHelper
 import com.talauncher.utils.UsageStatsHelper
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -16,7 +17,8 @@ import kotlinx.coroutines.launch
 class SettingsViewModel(
     private val appRepository: AppRepository,
     private val settingsRepository: SettingsRepository,
-    val usageStatsHelper: UsageStatsHelper
+    val usageStatsHelper: UsageStatsHelper,
+    val permissionsHelper: PermissionsHelper
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow(SettingsUiState())
