+++ app/src/main/java/com/talauncher/utils/PermissionsHelper.kt
@@ -1,0 +1,27 @@
+package com.talauncher.utils
+
+import android.app.AppOpsManager
+import android.content.Context
+import android.os.Build
+import android.os.Process
+
+class PermissionsHelper(private val context: Context) {
+
+    fun hasUsageStatsPermission(): Boolean {
+        val appOps = context.getSystemService(Context.APP_OPS_SERVICE) as AppOpsManager
+        val mode = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            appOps.unsafeCheckOpNoThrow(
+                AppOpsManager.OPSTR_GET_USAGE_STATS,
+                Process.myUid(),
+                context.packageName
+            )
+        } else {
+            appOps.checkOpNoThrow(
+                AppOpsManager.OPSTR_GET_USAGE_STATS,
+                Process.myUid(),
+                context.packageName
+            )
+        }
+        return mode == AppOpsManager.MODE_ALLOWED
+    }
+}
