+++ gradle.properties
@@ -3,4 +3,4 @@ android.useAndroidX=true
 android.enableJetifier=true
 kotlin.code.style=official
 android.nonTransitiveRClass=true
-org.gradle.java.home=C:\\Program Files\\Java\\jdk-24
+org.gradle.java.home=/root/.local/share/mise/installs/java/21.0.2
